first draft of speedo complete

+++ speedometer/frequency_measurement.h
@@ -12,7 +12,7 @@ class FrequencyMeasurer {
 		float maximum_gap_between_pulses;
 
 		bool make_measurement();
-		void set_frequency(unsigned float micros_gap);
+		void set_frequency(unsigned long micros_gap);
 
 	public:
 		FrequencyMeasurer(uint8_t input_pin, bool starts_at_high, float minimum_frequency);
